app/requirements.txt
databricks-sdk
streamlit
pandas
